+++ gi-gtk-declarative/src/GI/Gtk/Declarative.hs
@@ -65,6 +65,7 @@ import           GI.Gtk.Declarative.MenuModel  as Export
 import           GI.Gtk.Declarative.Patch      as Export
 import           GI.Gtk.Declarative.SingleWidget
                                                as Export
+import           GI.Gtk.Declarative.Slots      as Export
 import           GI.Gtk.Declarative.Widget     as Export
 import           GI.Gtk.Declarative.Widget.Conversions
                                                as Export